+++ flecsi/exec/fwd.hh
@@ -81,7 +81,8 @@ template<auto & Task,
     backend; each type involved documents the correspondence.
     User-defined classes can extend this mechanism by inheriting from
     \c params_tag or \c arg_tag.
-    If a parameter and its argument are each a (reference to a) \c std::vector
+    If a parameter and its argument are each (a reference to)
+    a \c std::vector, a \c std::optional,
     or a \c std::tuple or \c std::variant (of the same size), their elements
     are treated as
     separate parameters/arguments (with the unusual corollary that a